Feat: 회원 장바구니 선택물품 구매 넘어가기 및 레디스 트랜잭션 설정

+++ src/main/java/shop/s5g/shop/dto/cart/response/CartDetailInfoResponseDto.java
@@ -4,8 +4,8 @@ import java.math.BigDecimal;
 
 public record CartDetailInfoResponseDto(
     BigDecimal totalPrice,
-    int deliveryFee,
-    int freeShippingThreshold
+    long deliveryFee,
+    long freeShippingThreshold
 ) {
 
 }
